Moved DPE workflows:
* Unit tests were commented out, as we have some failures that seems unrelated so far
* Moved from requirements.txt to poetry and pyproject.toml
* Updated tox.ini
* Removed the global reference to tests.integration. in the integration tests to relative refs instead.
