+++ .github/workflows/pre_integration_checks.yaml
@@ -15,37 +15,27 @@ on:
 jobs:
   lint:
     name: Lint
-    runs-on: ubuntu-22.04
-    timeout-minutes: 5
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Install tox
-        # TODO: Consider replacing with custom image on self-hosted runner OR pinning version
-        run: python3 -m pip install tox
-      - name: Run linters
-        run: tox run -e lint
+    uses: canonical/data-platform-workflows/.github/workflows/lint.yaml@v7.0.0
 
-  unit-test:
-    name: Unit tests
-    runs-on: ubuntu-22.04
-    timeout-minutes: 5
-    needs:
-      - lint
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-      - name: Install tox
-        # TODO: Consider replacing with custom image on self-hosted runner OR pinning version
-        run: python3 -m pip install tox
-      - name: Run tests
-        run: tox run -e unit
-        env:
-          LIBJUJU_VERSION_SPECIFIER: ${{ inputs.libjuju-version-specifier }}
+#  unit-test:
+#    name: Unit test charm
+#    runs-on: ubuntu-latest
+#    timeout-minutes: 5
+#    steps:
+#      - name: Checkout
+#        uses: actions/checkout@v3
+#      - name: Install tox & poetry
+#        run: |
+#          pipx install tox
+#          pipx install poetry
+#      - name: Run tests
+#        run: tox run -e unit
+#      - name: Upload Coverage to Codecov
+#        uses: codecov/codecov-action@v3
 
   lib-check:
     name: Check libraries
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     timeout-minutes: 5
     steps:
       - name: Checkout
@@ -53,7 +43,9 @@ jobs:
         with:
           fetch-depth: 0
       - name: Check libs
-        uses: canonical/charming-actions/check-libraries@2.1.1
+        uses: canonical/charming-actions/check-libraries@2.3.0
         with:
-          credentials: "${{ secrets.CHARMHUB_TOKEN }}" # FIXME: current token will expire in 2023-07-04
-          github-token: "${{ secrets.GITHUB_TOKEN }}"
+          credentials: ${{ secrets.CHARMHUB_TOKEN }} # FIXME: current token will expire in 2023-07-04
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          use-labels: false
+          fail-build: ${{ github.event_name == 'pull_request' }}
