--- requirements.txt
@@ -1,19 +1,0 @@
-cosl
-boto3>="1.28.22"
-bcrypt==4.0.1
-cryptography==41.0.4
-jsonschema==4.19.1
-ops==2.7.0
-overrides==7.4.0
-pydantic==1.10.13
-requests==2.31.0
-ruamel.yaml==0.17.35
-tenacity==8.2.3
-urllib3==2.0.6
-pytest-mock==3.11.1
-parameterized==0.9.0
-shortuuid==1.0.11
-jproperties==2.1.1
-
-# Workaround to LIBJUJU issue
-protobuf==3.20
