+++ tests/integration/ha/test_ha.py
@@ -9,25 +9,7 @@ import time
 import pytest
 from pytest_operator.plugin import OpsTest
 
-from tests.integration.ha.continuous_writes import ContinuousWrites
-from tests.integration.ha.helpers import (
-    all_processes_down,
-    app_name,
-    assert_continuous_writes_consistency,
-    get_elected_cm_unit_id,
-    get_shards_by_index,
-    send_kill_signal_to_process,
-    update_restart_delay,
-)
-from tests.integration.ha.helpers_data import (
-    create_index,
-    default_doc,
-    delete_index,
-    index_doc,
-    search,
-)
-from tests.integration.ha.test_horizontal_scaling import IDLE_PERIOD
-from tests.integration.helpers import (
+from ..helpers import (
     APP_NAME,
     MODEL_CONFIG,
     SERIES,
@@ -40,8 +22,20 @@ from tests.integration.helpers import (
     get_reachable_unit_ips,
     is_up,
 )
-from tests.integration.helpers_deployments import wait_until
-from tests.integration.tls.test_tls import TLS_CERTIFICATES_APP_NAME
+from ..helpers_deployments import wait_until
+from ..tls.test_tls import TLS_CERTIFICATES_APP_NAME
+from .continuous_writes import ContinuousWrites
+from .helpers import (
+    all_processes_down,
+    app_name,
+    assert_continuous_writes_consistency,
+    get_elected_cm_unit_id,
+    get_shards_by_index,
+    send_kill_signal_to_process,
+    update_restart_delay,
+)
+from .helpers_data import create_index, default_doc, delete_index, index_doc, search
+from .test_horizontal_scaling import IDLE_PERIOD
 
 logger = logging.getLogger(__name__)
