+++ tests/integration/plugins/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2023 Canonical Ltd.
+# See LICENSE file for licensing details.
