+++ tests/integration/relations/test_opensearch_provider.py
@@ -11,8 +11,8 @@ import pytest
 from charms.opensearch.v0.constants_charm import ClientRelationName
 from pytest_operator.plugin import OpsTest
 
-from tests.integration.helpers import APP_NAME as OPENSEARCH_APP_NAME
-from tests.integration.helpers import (
+from ..helpers import APP_NAME as OPENSEARCH_APP_NAME
+from ..helpers import (
     MODEL_CONFIG,
     SERIES,
     get_application_unit_ids,
@@ -20,8 +20,8 @@ from tests.integration.helpers import (
     http_request,
     scale_application,
 )
-from tests.integration.helpers_deployments import wait_until
-from tests.integration.relations.opensearch_provider.helpers import (
+from ..helpers_deployments import wait_until
+from .helpers import (
     get_application_relation_data,
     ip_to_url,
     run_request,
