+++ tox.ini
@@ -3,117 +3,104 @@
 
 [tox]
 no_package = True
-skip_missing_interpreters = True
-env_list = format, lint, unit
-labels =
-    # Don't run this group in parallel, or with --keep-models because it creates a lot of local VMs.
-    integration = {charm, tls, ha, ha-networking, ha-multi-clusters, large-deployments, client, h-scaling, ha-storage}-integration
+env_list = lint, unit
 
 [vars]
 src_path = {tox_root}/src
 tests_path = {tox_root}/tests
-lib_path = {tox_root}/lib/charms/opensearch/v0
-all_path = {[vars]src_path} {[vars]lib_path} {[vars]tests_path}
+lib_path = {tox_root}/lib/charms/opensearch
+all_path = {[vars]src_path} {[vars]tests_path} {[vars]lib_path}
 
 [testenv]
 set_env =
-    PYTHONPATH = {tox_root}:{tox_root}/lib:{[vars]src_path}
-    PYTHONBREAKPOINT=ipdb.set_trace
+    PYTHONPATH = {[vars]src_path}:{tox_root}/lib
     PY_COLORS=1
     LIBJUJU_VERSION_SPECIFIER={env:LIBJUJU_VERSION_SPECIFIER:==3.2.3.0}
+allowlist_externals =
+    poetry
 
+[testenv:{build,pack-wrapper}]
+# Wrap `charmcraft pack`
 pass_env =
-    PYTHONPATH
-    CHARM_BUILD_DIR
-    MODEL_SETTINGS
-    LIBJUJU_VERSION_SPECIFIER
+    CRAFT_SHARED_CACHE
+allowlist_externals =
+    {[testenv]allowlist_externals}
+    charmcraft
+    mv
+commands_pre =
+    poetry export --only main,charm-libs --output requirements.txt
+commands =
+    build: charmcraft pack {posargs}
+commands_post =
+    mv requirements.txt requirements-last-build.txt
 
 [testenv:format]
 description = Apply coding style standards to code
-deps =
-    black
-    isort
+commands_pre =
+    poetry install --only format
 commands =
-    isort {[vars]all_path}
-    black {[vars]all_path}
+    poetry lock --no-update
+    poetry run isort {[vars]all_path}
+    poetry run black {[vars]all_path}
 
 [testenv:lint]
 description = Check code against coding style standards
-pass_env =
-    {[testenv]pass_env}
-deps =
-    black
-    flake8
-    flake8-docstrings
-    flake8-copyright
-    flake8-builtins
-    pyproject-flake8
-    pep8-naming
-    isort
-    codespell
+allowlist_externals =
+    {[testenv]allowlist_externals}
+    find
+commands_pre =
+    poetry install --only lint
 commands =
-    codespell {[vars]lib_path}
-    codespell {tox_root} --skip {tox_root}/.git --skip {tox_root}/.tox \
-      --skip {tox_root}/build --skip {tox_root}/lib --skip {tox_root}/venv \
-      --skip {tox_root}/.mypy_cache --skip {tox_root}/icon.svg
+    poetry check --lock
+    # uncomment the following line if this charm owns a lib
+    # codespell {[vars]lib_path}
+    poetry run codespell {[vars]all_path}
     # pflake8 wrapper supports config from pyproject.toml
-    pflake8 {[vars]all_path}
-    isort --check-only --diff {[vars]all_path}
-    black --check --diff {[vars]all_path}
+    poetry run pflake8 {[vars]all_path}
+    poetry run isort --check-only --diff {[vars]all_path}
+    poetry run black --check --diff {[vars]all_path}
+    find {[vars]all_path} -type f \( -name "*.sh" -o -name "*.bash" \) -exec poetry run shellcheck --color=always \{\} +
 
 [testenv:unit]
 description = Run unit tests
-deps =
-    pytest
-    pytest-asyncio
-    coverage[toml]
-    -r {tox_root}/requirements.txt
+set_env =
+    {[testenv]set_env}
+commands_pre =
+    poetry install --only main,charm-libs,unit
 commands =
-    coverage run --source={[vars]src_path} --source={[vars]lib_path} \
+    poetry run coverage run --source={[vars]src_path},{[vars]lib_path} \
         -m pytest -v --tb native -s {posargs} {[vars]tests_path}/unit
-    coverage report
+    poetry run coverage report
+    poetry run coverage xml
 
-[testenv:{charm, tls, client, ha-base, h-scaling, ha-storage, ha-networking, ha-multi-clusters, large-deployments, plugins}-integration]
+[testenv:integration]
 description = Run integration tests
+set_env =
+    {[testenv]set_env}
+    # Workaround for https://github.com/python-poetry/poetry/issues/6958
+    POETRY_INSTALLER_PARALLEL = false
 pass_env =
-    {[testenv]pass_env}
     CI
-    CI_PACKED_CHARMS
-    S3_ACCESS_KEY
-    S3_SECRET_KEY
-    S3_BUCKET
-    S3_REGION
-    S3_SERVER_URL
-    S3_CA_BUNDLE_PATH
-    # For AWS testing
-    AWS_ACCESS_KEY
-    AWS_SECRET_KEY
-    # For GCP testing
-    GCP_ACCESS_KEY
-    GCP_SECRET_KEY
-    # Generic secrets from CI:
+    GITHUB_OUTPUT
     SECRETS_FROM_GITHUB
-    TEST_NUM_APP_UNITS
-deps =
-    # This applies to libjuju, not Juju.
-    juju{env:LIBJUJU_VERSION_SPECIFIER}
-    opensearch-py
-    pytest
-    pytest-asyncio
-    pytest-operator
-    -r {tox_root}/requirements.txt
+allowlist_externals =
+    {[testenv:pack-wrapper]allowlist_externals}
+
+    # Set the testing host before starting the lxd cloud
+    sudo
+    sysctl
+
+commands_pre =
+    poetry install --only main,charm-libs,integration
+
+    # Set the testing host before starting the lxd cloud
+    sudo sysctl -w vm.max_map_count=262144 vm.swappiness=0 net.ipv4.tcp_retries2=5
+
+    {[testenv:pack-wrapper]commands_pre}
 commands =
-    charm: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/test_charm.py
-    tls: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/tls/test_tls.py
-    client: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/relations/opensearch_provider/test_opensearch_provider.py
-    ha-base: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_ha.py
-    # h-scaling must run on a machine with more than 2 cores
-    h-scaling: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_horizontal_scaling.py
-    ha-storage: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_storage.py
-    ha-networking: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_ha_networking.py
-    ha-multi-clusters: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_ha_multi_clusters.py
-    large-deployments: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/ha/test_large_deployments.py
-    plugins: pytest -v --tb native --log-cli-level=INFO -s {posargs} {[vars]tests_path}/integration/plugins/test_plugins.py
+    poetry run pytest -v --tb native --log-cli-level=INFO -s --ignore={[vars]tests_path}/unit/ {posargs}
+commands_post =
+    {[testenv:pack-wrapper]commands_post}
 
 [testenv:ha-backup-integration]
 description = Run integration tests
